+++ Flow/Changes.md
@@ -68,6 +68,7 @@
 ### Other changes
 - **XmlReader** now discards non-XML prefix from the streamed input, 
   effectively handling the BOM issue, and other such malformed inputs
+- **MutableOnce** now extends **Lockable**
 - **ValueConvertible**'s `.toValue` is no longer implicit
 - **UncertainBoolean** now extends **ValueConvertible**
 - Optimized `IntSet.from(IterableOnce)` implementation